MQTTImageRecognizer/TelegramBot/requirements.txt
paho-mqtt
python-telegram-bot==21.4
dotenv
